feat: ✨ Added repair mending items feature to xp pump

+++ src/main/java/net/p3pp3rf1y/sophisticatedbackpacks/upgrades/magnet/MagnetUpgradeWrapper.java
@@ -139,7 +139,7 @@ public class MagnetUpgradeWrapper extends UpgradeWrapperBase<MagnetUpgradeWrappe
 				}
 
 				if (amountToTransfer > amountAdded) {
-					world.addFreshEntity(new ExperienceOrbEntity(world, pos.x(), pos.y(), pos.z(), XpHelper.liquidToExperience(amountToTransfer - amountAdded)));
+					world.addFreshEntity(new ExperienceOrbEntity(world, pos.x(), pos.y(), pos.z(), (int) XpHelper.liquidToExperience(amountToTransfer - amountAdded)));
 				}
 				return true;
 			}
